[non-normative] Disable convolve range assertions
Per the linked bug report, it is possible for valid bitstreams to
hit the assertions on the range of 'sum' in the various convolve
functions.

On the master branch, we have added a new function which avoids this
situation. But in order to minimize changes on the av1-normative branch,
we instead simply remove the offending assertions.

BUG=aomedia:2074

Change-Id: If214b0b8325813ac0637b78a49c2cacf9bec8e9a