allow_warp: rectify mismatch between spec and code
As per the spec, we must compare the scales at higher precision in this
function to check if there is a rescale.

So, we use the av1_is_scaled() function that looks at the fixed point
scales.

Cherry-pick of 14b933f9c14d08e3e79713bff0e431e4456b7f24

BUG=aomedia:2191

Change-Id: I9276dc6f54fe01deb98549ac4ebcc25fae462cf2