data-driven-tx: ddtx for intra
Introduce KLT, GFT, and LGT for intra blocks. Three 2D transforms
are added for intra blocks: DDTX1_DCT, DCT_DDTX1, DDTX1_DDTX1.
The transforms for intra blocks are common for all modes (not
mode-dependent).

The is_inter symbol is passed to the 1D transform level, so the
signature of 1D transforms are changed. In entropy coding, an
additional symbol use_ddtx_intra is used to transmit the
transform type.

Change-Id: Ib1e867d10c808db4832a096092ce753e5b5ef961