3-way partitions: update buffer offset logic.
In preparation for enabling 3-way partitions of 16x16 blocks.

Design: Only the 3rd block in a 3-way partition of a 16x16 block is a
chroma reference, and the corresponding chroma block will be of size
8x8.

Update the logic to offset the buffer for Chroma in such a case, in
various parts of the codebase -- offset will be 3 for the 3rd sub-block.

Change-Id: I7a98d0a71ad02d760d2c2a4a18c40365b3e4cabe