Use is_mi_coded map to determine has_top_right in intra coding
For AV2 baseline, no result changed should be expected.
When 3WAY_PARTITIONS tool is enabled, a slight performance change
is introduced because of a corner case: for the 2nd block of a
vertical 3way partition, for example, 8x16 in a 3-way partitioned
16x16, if 4x4 tx is applied, the new algorithm will determine that
the top-right block of the 8x16 does have its top-right reference
available if the above row is available.

Change-Id: I27f6273914b9f8fde2e15559d45be52a76ab6096