Add the option to use flex_mvres at sb level
This CL introduces the CONFIG_SB_FLEX_MVRES flag under
CONFIG_FLEX_MVRES, which moves the mvres decision to the superblock
level instead of partition block level. CONFIG_SB_FLEX_MVRES is on by
default when CONFIG_FLEX_MVRES is on, but it can be turned off in
av1/common/mv.h.

Change-Id: I106e88ecf4d764c415f16c8288bec96d81f6f46a