Command-line option for automatic fixed QP offsets
The new command-line option is: --use-fixed-qp-offsets=1, which will
trigger the use of fixed QP offsets. Offsets are either:
- User specified when --fixed-qp-offsets=a,b,c,d,e is also given, OR
- Automatic when --fixed-qp-offsets are NOT given.

Implementation mimics the QP offsets used by default 1-pass encode for
Alt-ref and internal ARFs. And the QP offset for keyframes is chosen
empirically.

Coding performance is close to baseline (+0.2).

(Baseline: --passes=1 --lag-in-frames=25 --kf-min-dist=64
--kf-max-dist=64 --min-gf-interval=16 --max-gf-interval=16
--gf-min-pyr-height=4 --gf-max-pyr-height=4 --limit=128 --end-usage=q
--cq-level=x
Test: Additionally '--use-fixed-qp-offsets=1')

Change-Id: I5430baa63607ff0ca34d54765aec4f8ced5089a5