Add global MV to singleref refstack when possible.
This change adds the frame global motion vector to the reference stack
whenever there is empty space for it.

This brings the performance of NEW_INTER_MODES to -0.216% on lowres150.
Encoding speed change was not measured.

Since this patch was developed, changes to the baseline have reduced NEW_INTER_MODES performance.

Change-Id: I7016d9a3882465a0b40fb9734155bf5d9a79b234