A fix to generalize has_top_right() and has_bottom_left()
Directly use mi_row_chroma_base and mi_col_chroma_base in
mbmi->chroma_ref_info to determine whether top-right and
bottom-left reference is inside the superblock or not. The previous
algorithm using bsize to figure out the aligned base location does
not work in the case when ext-recur-partitions and ext-partitions
are both enabled.

Change-Id: Iba2cb99980e3822e4c28cdc9ac9a8ca6f34b4997