Reduce redundant partition searches in the new partition tool
Check whether there is a counterpart pc_tree node with the same size
and the same neighboring context (by ensuring the ancester nodes all
have index 0) at the same location but from a different partition
path, and if yes directly copy the RDO decision made for the
counterpart.

It reduces the encoder complexity of EXT_RECUR_PARTITIONS +
EXT_PARTITIONS from 126x to 83x of its baseline.

Change-Id: I0cbe7591301bff9aa96f4f2de8d8f9af58cc4864