Avoid signed overflow for quantizer > 32767
When ext_quant is enabled, the quantizer value can be larger than 32767.
when the quantizer is > 32767, the variable 'l' in the invert_quant(),
which is equal to trunc(log2(d)) can be as large as 15. This would cause
the '1 << (16 + 15)' overflow a signed integer(i.e. get a negative value).
This would cause wrong result for later steps. Adding a 'u' postfix
will force an unsigned representation.

Change-Id: I4786dbe4feb53fcd3c03d428a2292ea80f03392c