Add an experiment to send a base delta_q for dc
This experiment (CONFIG_DELTA_DCQUANT) removes the
dc quant tables and instead represents the dc quant
values using a delta-q offset from the ac quant tables.

With a default value of 8 for the new offset, BDRATE in
constant q mode on 33 frames at cpu-used 0, are as follows:
lowres2: +0.235% Av PSNR, -1.257% SSIM)
midres2: -0.278% Av PSNR, -1.485% SSIM)
hdres2: -0.234% Av PSNR, -1.273% SSIM)

There are sevaral empirical parameters in the encoder
including the dead-zone size for the quantizers that
need to be tuned for optimum results.

Change-Id: I0146cca5dfe1ea774f7116a8c49f7b61ac9fbb6d