Change the behavior of EXTQUANT / DELTA_DCQUANT
Changes the behavior of the two experiments so that
DELTA_DCQUANT can be tested independently for now.
A compile rule is added so that EXTQUANT automatically
turns on DELTA_DCQUANT, because EXTQUANT assumes no
dc tables are needed.

When the individual experiments are finalized the experiments
may be merged into one.

Change-Id: If82cb040dc956d2eb52164b28c00fa249f758720