Disable printing per-frame stats in "--quiet" mode
Earlier, report_stats() function was printing per-frame stats even when
"global.quiet == 1", which was incorrect. This also means a lot of log
is printed when running unit tests (nothing was printed before).

We fix this by passing 'global.quiet' value to 'cpi' through
'base.init_flags'. This is similar to how 'global.show_psnr' is passed
to 'cpi'.

Change-Id: Ic1fb556f32efe50705a4c71649d9ff0d2ad27fad