Update the ref frame stacks according to subgop cfg
This will use the frame types from the cfg rather than the frame update
types. Currently, this will organize the frames into the stacks in the
same way as the baseline. Any changes to this scheme can be done in
update_ref_frame_map_gopcfg. This currently does not work with TPL.

Change-Id: Idf96b235ce1e4276cc5b0c69ac59b52e3818d791