Fix calculation of arf_src_offset
For asymmetric subgop cfg, we generalize the calculation of
arf_src_offset by recording the last shown frame (instead for
a perfect binary pyramid, the calculation won't account for the
case that there are multiple arfs at the same level in a mini
interval). Also for the case that leaf frames are not all at the
bottom level, properly mark the last level frames.

Change-Id: I811880efad5d5c02a4339ee2d00621518d9d397d