Modify quantizer for CONFIG_DSPL_RESIDUAL
Downsampling is a lossy operation which significantly reduces
information content in the the transform coefficients. To offset this
loss, it is important to modify the quantizer to be less aggressive when
downsampling.

This change makes this modification. For the encoder, a new method
av1_setup_dspl_quantizer() is introduced to modify the Y-plane's
quantizer. Note that this method merely swaps pointers to already built
quantizers, the quantizers are not re-built.

For the decoder, the dequantizers are only built for the required
qindices. Thus dequantizers are built for both the downsampling options.

The qindex offset is currently static and is set by the
av1_calc_dspl_qindex() method.

Change-Id: I02dce5d69391aae0a5b16071e8074931393569fd