LOOP_RESTORE_CNN: bugfix for search_cnn() function
Earlier, the case when CNN should be disabled wasn't handled correctly.
So, it was possible for CNN to end up being used when it wasn't supposed
to be used, if bits_cnn < bits_none.

We fix this by setting appropriate vars and existing early for this
case.

Comparison of gains after this patch:
* Random Access:
                CNN_RESTORATION   LOOP_RESTORE_CNN
lowres2           -1.797            -1.541
midres2           -2.039            -1.966
objective-1-fast  -1.846            -1.826

* All Intra:
                CNN_RESTORATION   LOOP_RESTORE_CNN
lowres2           -4.701            -3.700
midres2           -5.083            -3.982
objective-1-fast  -4.865            -3.753

Change-Id: I6319c7b72226c2c37ad7cd0463a85662b1c6c81d