Bugfix: low delay with pyramid (adaptive QP case)
Frames are assigned qp based on their pyramid levels as follows:
- Starting with a quality of pyramid level 1, determined based on some stats
- Then selecting a value between that and cq_level, depending on
pyramid level.

This was working fine for high delay case (out-of-order coding), as ARF
frame comes before any frames at lower levels of the pyramid. So, one
can store qp for ARF in 'arf_q' and use that for later frames.

However, for low-delay case, frames are coded in display order. So,
'arf_q' was used uninitialized (had value 0).

To correct this for low delay case, now initialize this variable for the
first frame that is not a bottom leaf frame.

Also, renamed this variable to more appropriate name 'level1_qp'.
And added some checks to avoid uninitialized values in the future.

Coding gain for low delay default vs --subgop-config-str=ld:
                    Before    After
lowres2:            +1.026    -3.021
midres2:            +3.295    -3.053
objective-1-fast:   +3.515    -4.404

Change-Id: Ic57091de14acf2cb9881673301aa53e054611f84