Pass a boolean as the has_val argument to ARG_DEF
The third argument of the ARG_DEF macro is the has_val field of the
arg_def struct. It is not the default value of the command-line option.
So change 4 and 128 to 1 (true).

Note: These two ARG_DEF macro calls were added in
commit 9543c705874fef1d11d4ef539170fd89c0197929.

Change-Id: I06a451e5cfc591796b05f7ba173f3c95e2491993
(cherry picked from commit 432f0306881fde3d7106c4a137d2a5acf8081035)