Fix reference filter for RST_MERGECOEFFS.
We need to keep track of the most recent reference filter at the start
of the merge stack to get accurate bit counts. To this end, a ref_wiener
field was added to RstUnitSnapshot. merge_sse and merge_bits have also
been added to complete the structure and simplify stack updates.

Change-Id: I22db53824bec25e307206a75f6bd8c498246d069