Initialize the tx_skip array according to chroma ref offsets
When encoding a block, the part in cm->tx_skip array representing
the area will be reset as all 0. If several chroma blocks are
combined as one chroma coding block, since it is performed when the
last block is reached, we need to use the offset information to
reset the whole area. This resolves an assertion error in
FLEX_PARTITION.

Change-Id: Ia7d0b6e8dcee9d807e9eb6b2723da7db1d5191d8