Derive reference frame mapping on decoder side
This avoids sending the mapping explicitly when order_hints are enabled.
It includes sending 1 bit at the frame level to indicate if we are in
realtime mode (realtime uses a different mapping scheme),
2 bits at the frame level to indicate information
about the pyramid level of the current frame, and removing all of the
reference map index signaling. 0.38% gain on lowres2.

Change-Id: I965b328c96ffb8846414b4b931944c2abad27c6a