Fix for race condition with SDP
There is one nightly test failure regarding to data race when SDP is on. The problem is because that tile level struct cpi->td is used in function pick_intra_angle_routine_sbuv. To solve the problem, instead of using tile level struct, block level struct x is used in this function.

Change-Id: I7f45c74e929329952558c9a3eb80dc23e5267cf1