Refactor av1_compute_subpel_gradients and some cleanups
The function for gradients computation is refactored into two
functions, one for generating the predicted block and the other for
obtaining subpel gradients. This is done to clean up code structure
in interpolation based gradients, and this change will be needed
for a later patch where computations of gradient and least squares
are partially combined.

Change-Id: Ic6818ec79aed448a324e08dfacdbeab8b28d5de6