Code cleanup for CONFIG_ORIP and CONFIG_SDP
For vertical and horizontal intra prediction mode, ORIP is only applied to luma component. So there is no need to pass the xd->tree_type to function av1_signal_orip_for_horver_modes.
In addition to that, this commit changes the way to #if..#elif..#endif when CONFIG_SDP and CONFIG_ORIP interact.

This commit doesn't change the bitstream.

Change-Id: I656d578d5184d45c37ac5d033e5571b1c102b735