Bugfix motion field estimation
Based on the logic on encoder side and the syntax on the AV1 spec,
the motion field estimation should be turned on or off based on
the control flag allow_ref_frame_mvs instead of always enable it
on the decoder. If encoder's running time flag --enable-ref-frame-mvs
is set to 0 (disable), then the motion field estimation on encoder is
disabled, however the motion field estimation on decoder is still
enabled due to this bug, which will cause the inconsistent between
encoder and decoder.

Change-Id: I339cbb747b967dafa73e6f2b5cd192acfbe595aa