Bugfix SMVP improvement when MF estimation is off
The mismatching reported in BUG=aomedia:3094 is because SMVP uses
some variabls which value are derived in motion field estimation.
However, there is a bug in av1/av2 code when encoder's running time
flag --enable-ref-frame-mvs is 0 (disable). If --enable-ref-frame-mvs
is false, the motion field estimation on encoder is disabled, however
the motion field estimation on decoder is still enabled due to av1/av2
bug. This will cause inconsistent results between encoder and decoder.
After add the control flag allow_ref_frame_mvs to control the on and off
of motion field estimation on decoder, it can resolve the mismatching.

Meanwhile, some code was added in SMVP improvement to handle the case
when --enable-ref-frame-mvs is false.

BUG=aomedia:3094

Change-Id: I2fa8d68a03da2b42f46a49a8f460e53871b51fda