Modify assertion check for max-partition-size with CONFIG_SDP
When CONFIG_SDP is enabled, the minimum value of max-partition-size is changed from 4 to 8. When max-partition-size is equal to 4, the chroma block size must be 2x2 for yuv420 format since there are no other options. However, 2x2 chroma block size is disallowed when SDP is enabled.  In this case, the decoder will crash since 2x2 chroma block is not supported when SDP is on. So, the max-partition-size must be 8 or greater than 8 when SDP is on.

BUG=aomedia:3087

Change-Id: I0b9ac8bc129af48b6cba156606defce4b8427d46