Local tests for spherical video
This is the local test code for the spherical video project. This
experimental code is for reference only and is not intended for
production use.

The y4m reading code is empty since it is not from Google. You have to
implement it by yourself following the interface and comments

The tools include writing motion search raw data (coordiante, pixel,
filter kernel, filter result, SAD, SAD map) to txt/csv files,
calculating statistical data,
highlighting the reference block, and motion search for blocks evenly
distributed across the frame.

Change-Id: I75c7abdbbff0fb461ffc2457ca9e7101f2fde6f2