Take away one bit from pyramid signaling in newref
Currently the ranking mechasnism only uses information on
whether a frame is at the lowest pyramid level or not.
So simplify the syntax to remove the second bit that was
being signaled to indicate highest pyramid level.
Later when we move entirely to using a qindex and distance
based ranking scheme this bit can also go away.

Change-Id: I6bf22460bd87d50631f06aa4924883e1d89d65bf