rtc: Fix is_key_frame setting in variance partition
On resized frame, or if LAST is NULL, the frame is treated
as key frame in variance partitioning (lines ~1406),
but in fill_variance_tree_leaves() the is_key_frame is
set again and does not include logic from lines ~1406.

This leads to worse quality on resized frame, and
causes a crash for resizing with 10bit input.

Fix is to pass the is_key_frame into the
fill_variance_tree_leaves().

Change-Id: Ibe4162de54cf31739667796a9666880293bd54ad
(cherry picked from commit a45d28c824e53931d72a1cfb02642611cc3d2321)