Fix buffer overrun in dist_wtd_convolve_2d_horiz_neon
When introducing the 6-tap specialization for
dist_wtd_convolve_2d_vert_neon[1], we attempted to reduce the number
of rows processed in the horizontal convolution (by 2) if the
subsequent vertical convolution would be using a 6-tap filter instead
of an 8-tap filter. This logic was faulty and meant we ended up
accessing memory outside of the (padded) source buffer - since the
horizontal convolution processes 4 rows of data per iteration, when
the number of rows to be processed is not necessarily a multiple of
4.

This patch restores the previous logic and src pointer starting
position for dist_wtd_convolve_2d_horiz_neon.

[1] Commit hash: be1d80024928684b1c9eebc648ed92d8ea70d166

Bug: aomedia:3367
Change-Id: Id11d4ae7c123957d3336b61f4d4cc8da09131b68