Correct buffer fullness Q adjustment.
Fix anomaly in adjustment of active worst quality for RTC.

The real time code adjusts the active worst Q allowed for each
frame based on buffer fullness. The baseline value is derived
from the ambient Q for recent frames. Where there has
been overspend and the buffer has run empty, the allowed worst
Q is increased. Conversely, where we have underspend and the
buffer fullness has risen above the target optimal level, we apply
a downward pressure on Q

This patch fixes an anomaly around transition between these two.

Adjustments up were applied to the measured ambient Q while
adjustments down were applied to the ambient Q * 5/4. This led
to the situation where as the buffer fullness rose above the optimal
level we would transition from (ambient Q +adjustment) to
((ambient Q *5/4) - adjustment). This meant that the value would
initially jump up higher for a time when one would expect it to start
falling.

The main observed impact of this patch is on the RTC screen set.

Bdrate improvements are as follows  (opsnr , ssim, vmaf) :-

Screen data set
Speed 10		-2.530, -2.174, -2.187. (speedup 0.769%)
Speed 8		-4.282, -3.816, -3.482, (speedup 3.432%)
Speed 6		-4.108, -3.730, -3.370, (speedup 3.060%)

*** NOTE ***
This patch was previously  reverted because of a bug when --lossless
is set, that gave rise to a negative value for ambient_qp. This should
now be resolved so the patch is being re-introduced. The metrics
gains may be a little different because of other changes in the
meantime.

STATS_CHANGED

Change-Id: I5dcdd51e8a2d43925e7a8c8299fbc0a93fdd8c16