Disable GM for frames at 0 distance
In some rare cases, we may end up wanting to search global motion
where the source and ref frames are at the same temporal instant.
An example would be coding a "real" frame against a temporally
filtered version of the same frame.

In practice this only arises in exceptional conditions, but the code
still needs to handle this case. However, in all such cases we expect
that the best global motion model will be no motion (IDENTITY).

By explicitly disabling global motion when the temporal distance is 0,
we can simplify the logic in a few places without changing results

Change-Id: I718eb6fe927dde1c6a01fc4bdb150189195120bc