thread_loop: lock/unlock mutex outside thread loop
It is not necessary to release and reacquire worker->impl_->mutex_
between iterations of the thread loop. We can acquire
worker->impl_->mutex_ before entering the thread loop and release
worker->impl_->mutex_ after exiting the thread loop.

Once we do this, we can exit the thread loop with a break statement and
get rid of the `done` boolean flag.

Finally, since worker->status_ has only three possible values, if it is
not OK and WORK, then it must be NOT_OK. So we can replace the last
else-if with an else. Also, if worker->status_ is assigned an invalid
value, the original code will enter an infinite loop. The else branch in
the new code will treat the invalid status as NOT_OK and exit the thread
loop.

Change-Id: I34f0bad43fe4f74dfc03813b59f64cc97b4cb6b9