av1_round_shift_rect_array_32_neon: fix undefined behavior
(left bitshift by negative amount)

Though av1_round_shift_rect_array_32_neon() has a check for `bit > 0`,
it unconditionally does:
  const int32x4_t rnding = vdupq_n_s32(1 << (bit - 1));
which becomes `1 << -1` [1].

The 'bit > 0' check and the setup done for it is unnecessary since this
function is always called with bit=0, val=NewSqrt2Bits. This change
removes those two parameters.

[1] https://www.open-std.org/jtc1/sc22/wg14/www/docs/n1256.pdf
Section 6.5.7
"If the value of the right operand is negative or is greater than or
equal to the width of the promoted left operand, the behavior is
undefined."

Bug: b:266719111
Change-Id: Ic9a2d82ef34f2a9b5bda57fe15b54019e87b582a
(cherry picked from commit f3e932d4ffe12c680d557635205497d347e5aef1)