Add kUnknownRefFrame to Q Mode API.
This allows the TPL stats to signal that a block was inter coded
but the reference frame isn't known.

Also added `TplFrameStats::ref_frame_indices`, listing all the
reference frames which the encoder used when generating TPL stats
for a particular frame. When a `TplBlockStats::ref_frame_index`
contains `kUnknownRefFrame`, it's known at least that the reference
was one of those frames.

For now, `GetRefCodingIdxList` handles `kUnknownRefFrame` by
assuming that the reference was the first frame in
`TplFrameStats::ref_frame_indices` -- which preserves the
existing behavior -- and added a test for this case.

Bug: b/266007381
Change-Id: Ib1c58582c2565e2c6db36cf0bced01593d8a7909