Refactor weight calculation in highbd_apply_temporal_filter()
In highbd_apply_temporal_filter(), few of the variables related
to weight calculation were calculated for every pixel though
they are constant at sub-block level. In this CL, the
relevant calculations are done beforehand and the same are
consumed based on the sub-block index in both SSE2 and AVX2
variants of the function.

For 'good' encoding,
               Instruction Count       BD-Rate Loss(%)
cpu Resolution   Reduction(%)    avg.psnr  ovr.psnr  ssim
 5     LOWRES2      0.095        0.000     0.000     0.000
 5     MIDRES2      0.114        0.000     0.000     0.000
 5      HDRES2      0.130        0.000     0.000     0.000
 6     LOWRES2      0.051        0.000     0.000     0.000
 6     MIDRES2      0.099        0.000     0.000     0.000
 6      HDRES2      0.113        0.000     0.000     0.000

Change-Id: I566239e074a3e80e63166fb9ef5c6025d261bb80