Fix bug when --lossless selected
Incorrect behavior when --lossless selected.

The lossless command line flag is only valid for Aq mode 0 and
forces both the min and max quantizer to 0.

However, this special case code intended to give a boost to Q a few
frames after a poor quality key frame breaks the correct behavior
of lossless mode and causes a few frames to be coded at Q=5
instead of Q=0.

This special case is likely no longer very relevant but simply deleting
it causes some metrics regression.  It is likely that this regression
can e resolved in a subsequent patch and the code can then be
removed but this patch just adds an additional bounds check to make
sure Q is not forced above the allowed maximum.

Change-Id: If74e1bc345119b5a55fb2c8fc6f6e173e6e860f6