Reset avg Q accumulators for --lossless.
The --lossless command causes the min and max Q to be forced
to 0 even if not set to 0 on the command line.

This, however, occurs after the primary rate control has been
initialized in the processing of the extra config parameters.
The code added here ensures that the resetting of the min
and max Q is propagated correctly to the average Q rate control
accumulators.

Change-Id: Ic6dac7cb299258b713ca3101417d82b0823a1e50