highbd_inv_txfm_neon: fix undefined behavior
by callers of round_shift_4x4. This avoids preparing unused values when
shift == 0 and fixes sanitizer warnings in that case due to a left shift
of a negative value (1 << -1), which has undefined behavior [1]. Rather
than calculate the rounding value directly, vshlq_s32 is replaced with
vrshlq_s32.

[1] https://www.open-std.org/jtc1/sc22/wg14/www/docs/n1256.pdf
Section 6.5.7
"If the value of the right operand is negative or is greater than or
equal to the width of the promoted left operand, the behavior is
undefined."

Bug: b:266719111
Change-Id: I07594dabeee0b8cc4415e99fa9e50409362eaf55
(cherry picked from commit 73013d3ff036030deb627c940c44d2eda1fa345d)