Set max and min partition size in recode loop
In `av1_rd_pick_parition`, the maximum partition size is not enforced if
it enters a recode loop. When this happens, `encode_sb` not might be
called for some blocks, as `should_do_dry_run_encode_for_current_block`
returns false when the `bsize` is greater than the maximum block size.
This eventually leads to mismatch.

This commit enforced the maximum bsize constraint by calling
`av1_prune_partitions_by_max_min_bsize` if it enters recode loop.

STATS_CHANGED is theoretically possible, but none was observed in the
test clips.

Change-Id: Ic14c9f48a036c062bad0f6ac3345542802d99644