Specialize av1_dist_wtd_convolve_y_neon for 6-tap filter
av1_dist_wtd_convolve_y_neon can be called with 4-, 6- and 8-tap
filters, with 4- and 6-tap filters being padded out with 0s to enable
re-use of the same 8-tap code path. This is inefficient as we end up
spending a lot of time multiplying by, and adding, 0 - especially
since the most common filter size used with this function is 6-taps.

This patch adds a av1_dist_wtd_convolve_y_neon code path specialized
for 6-tap filters. This new path is used for both 4-tap and 6-tap
filters to reduce the amount of redundant work.

Change-Id: I730d6d1ff2e6316aedbc123c24bc26dc5a203dd1