Tune global motion segmentation
When evaluating global motion, we divide the frame into 32x32 pixel
segments so that we can evaluate the impact of global motion only on
the regions where it is likely to be used.

This helps us make better decisions about whether to use global motion.
However, if we can't find enough useful segments then we fall back
to considering the impact on the whole frame.

The threshold for "enough useful segments" was set proportionally to
the frame size, but the number of feature points used to construct
the map has a fixed maximum. This meant that large frames had a very
hard time satisfying this criterion, causing a fallback to looking
at the whole frame, where the model would probably be rejected.

So we change to using an absolute number of segments as the threshold.
Setting the threshold to 48 segments seems to give much more consistent
behaviour across sizes - it allows global motion to be used much more
on large frames without harming small frames.

Note that, as well as allowing global motion to be used more often,
this change also makes refinement much cheaper on average, leading
to a significant time saving at speeds 1 and 2

      |      lowres2       |      midres2       |       hdres2
Speed |  BDRATE | Enc time |  BDRATE | Enc Time |  BDRATE | Enc Time
------+---------+----------+---------+----------+---------+---------
  1   | -0.018% | -0.015%  | -0.146% | -0.471%  | -0.048% | -0.981%
  2   | +0.008% | +0.028%  | -0.124% | -0.232%  | -0.060% | -0.866%
  3   | -0.009% | +0.006%  | -0.105% | +0.863%  | -0.104% | +1.731%
  4   | +0.013% | -0.021%  | -0.160% | +0.999%  | -0.127% | +2.384%

STATS_CHANGED

Change-Id: I14c4b6e9393d6a8b188c9709a091a3375833e581