Disable global motion in lightfield mode
Lightfield videos expect all frame headers to be identical (see
logic in tests/lightfield_test.sh). This is incompatible with
global motion, which stores the per-frame motion models in the
frame header.

This test previously worked because global motion was automatically
disabled at speed 3, which is used by examples/lightfield_encoder.c.
But we recently enabled global motion at this speed setting, which
caused the test to intermittently fail.

To fix this, add logic which disables global motion when large scale
tile mode is enabled.

Change-Id: Ia40cc347d7a381963189d17accfe5065ab1497b5