Add Neon implementation of aom_convolve8_horiz
Add an Armv8.0 Neon path for aom_convolve8_horiz, and a corresponding
set of unit tests.

A Neon implementation of aom_convolve8_vert will be added in a
subsequent patch.

The Neon code added in this change is mostly a port of the Armv8.0
Neon implementation of the 'vpx_convolve8_horiz_neon' function in
libvpx.[1]

[1] https://github.com/webmproject/libvpx/blob/main/vpx_dsp/arm/vpx_convolve8_neon.c

Change-Id: I038508fb2a1f4db4eaa9119a210fc8c206665a4a