Clean up parse_layer_options_from_string()
Use the original memory allocation failure check but move it to the
right place. See
https://aomedia-review.googlesource.com/c/aom/+/119462
https://aomedia-review.googlesource.com/c/aom/+/158102

Set the error code when breaking early from the for loop rather than
detect an early break and set the error code after the for loop.

Check the return value of parse_layer_options_from_string().

Bug: aomedia:3404
Change-Id: I5ef325b1c067f18750568a7e78e9485c7dfe17a2