Tweak some static analysis suppressions
A follow-up to commit 4c5a1df0cc06b3b101786454070544641a44cebb.

If whitespace changes are ignored, there are only two changes to
third_party/fastfeat/nonmax.c in that commit, and we tweak them as
follows.

1. Change:
    if(pos.y > 0)
      if (row_start[pos.y - 1] != -1)
to:
    if(pos.y > 0 && row_start[pos.y - 1] != -1)

This is closer to the original code, which is:
    if(pos.y != 0 && row_start[pos.y - 1] != -1)

2. Remove the "if(pos.y >= 0)" check after the comment
"Check below (if there is anything below)". The scan-build in LLVM 14
does not need this check. Again this is closer to the original code.

After these tweaks, the description of
commit 06540eb9ad297a93b413c38041a95cceb6524c06 in
third_party/fastfeat/README.libaom also covers change #1, so I just
change the word "comparison" to "comparisons" (plural).

Change-Id: Ic6299a95f284dcb84379caf663518e6ab34faa02