Refactor rounding shims in averaging Neon convolution functions
In many cases when computing a sum using Neon, adding a constant shim
can avoid a 'complex' rounding right shift (which is usually slower
than a 'simple' truncating right shift.) Such a shim was added when
averaging in the Neon compound convolution functions. However, since
we are now using a 'complex' saturating-convert-to-unsigned right
shift (vqshrun_n_s16) in this sequence anyway, there's no extra cost
to using the rounding-saturating-convert-to-unsigned right shift
instruction (vqrshrun_n_s16.)

This patch removes the rounding constant shims in favour of using the
rounding-saturating-convert-to-unsigned right shift instruction to
compute the average in Neon compound convolution functions. It also
makes a number of cosmetic changes to make variable names consistent
the C reference and other Neon implementations.

Change-Id: I10efbcc687cbbaa1ed1aed09e4d57b3de515cf70