Specialize non-averaging dist_wtd_convolve_x_neon
Currently the code paths for selecting the appropriate type of averaging
have branches inside of the inner loops of the function - which is
rather inefficient. Having a specialised function for each of the three
paths (non-averaging, basic averaging and compound averaging) will allow
the branch to be made only once at the start of the function.
This patch specialises av1_dist_wtd_convolve_x_neon for averaging and
non-averaging paths. A following patch will further specialise the
averaging path between basic averaging and compound averaging.

This change and the one following bring up to 10% speedup on Clang and
up to 20% on GCC.

Change-Id: I81c193bfc37dc9e81c202e18a21a0c1970e9855f