Specialize non-averaging dist_wtd_convolve_2d_neon
Currently the code paths for selecting the appropriate type of averaging
have branches inside of the inner loops of the function - which is
rather inefficient. Having a specialised function for each of the three
paths (non-averaging, basic averaging and compound averaging) will allow
the branch to be made only once at the start of the function.  This
patch specialises the vertical pass of av1_dist_wtd_convolve_2d_neon for
averaging and non-averaging paths. A following patch will further
specialise the averaging path between basic averaging and compound
averaging.

This change and the one following bring up to 10% speedup on both Clang
and GCC.

Change-Id: Ib59eded0afb682258ce1ce5894f9c5418ab6427c