cmake: add AOM_ARCH_AARCH64
This will allow identifying Windows Visual Studio targets as aarch64;
the Microsoft compiler does not define __aarch64__.

An alternative would be to define this in the code, checking for
_M_ARM64 or _M_ARM64EC. For now we'll use the existing AOM_ARCH_*
system. For compatibility AOM_ARCH_ARM will continue to be defined to 1
in this case.

This is similar to the change in libvpx:
33aba6ecc configure: add aarch64 to ARCH_LIST

Bug: b/277255390
Change-Id: I76075257ac0fda411294480280605a38f67448c7