rtc-screen; Adjust condition for idtx mode with color
For idtx mode in nonrd pickmode screen content:
disallow the usage of the idtx mode on certain conditons
to avoid artifact (observed in sc_web_browsing clip).

For blocks with high motion and spatial variance below threshold:
if rdcost for y channel is exactly zero and nonzero for color,
better to not risk selecting idtx, as it can lead to color
artifacts.

And remove unneeded comment.

Neutral/small change on rtc screen, speed 10:
0.23% bdrate loss, -0.3% IC change.

Change-Id: I4815e10de7838e48a744f3fd8eebb40da3cb983e