Add Armv8.6 USDOT Neon implementation of aom_convolve8_horiz
Add an Armv8.6 Neon USDOT (dot-product) path for aom_convolve8_horiz.
The existing implementations are retained and used on target CPUs that
do not implement the USDOT instruction. The availability of the USDOT
instruction is indicated by the feature macro
__ARM_FEATURE_MATMUL_INT8.

An Armv8.6 USDOT Neon implementation of aom_convolve8_vert will be
added in a subsequent patch.

The Neon code added in this change is mostly a port of the Armv8.6
Neon implementation of the 'vpx_convolve8_horiz_neon' function in
libvpx.[1]

[1] https://github.com/webmproject/libvpx/blob/main/vpx_dsp/arm/vpx_convolve8_neon.c

Change-Id: Idaf7f74a315a88a105cb29ae725f1380d8127701