Add Armv8.6 USDOT Neon implementation of aom_convolve8_vert
Add an Armv8.6 Neon USDOT (dot-product) path for aom_convolve8_vert.
The existing implementations are retained and used on target CPUs that
do not implement the USDOT instruction. The availability of the USDOT
instruction is indicated by the feature macro
__ARM_FEATURE_MATMUL_INT8.

The Neon code added in this change is mostly a port of the Armv8.6
Neon implementation of the 'vpx_convolve8_vert_neon' function in
libvpx.[1]

[1] https://github.com/webmproject/libvpx/blob/main/vpx_dsp/arm/vpx_convolve8_neon.c

Change-Id: I4b98cf72556ac665bac68a36ae19d853af5a2e28