Move encoder-only global motion functions
av1_frame_error and related functions are only used for non-realtime
encoding. So move them to files which a) are under av1/encoder, and b)
are excluded from realtime-only builds. This way, the functions are
automatically skipped over in configurations where they are not needed.

Also add av1_calc_highbd_frame_error to RTCD system, so that we can write
SIMD acceleration for it in future.

Change-Id: I9b8c257cc0e15998ecd14bd320e3ae5b87bd157a