Fix C vs intrinsic mismatch of aom_hadamard_32x32() function
This CL resolves the mismatch between C and intrinsic implementation
of aom_hadamard_32x32 function. The mismatch was due to integer
overflow during the addition operation in the intrinsic functions.
Specifically, the addition in the intrinsic function was performed
at the 16-bit level, while the calculation of a0 + a1 resulted in
a 17-bit value.

This code change addresses the problem by performing
the addition at the 32-bit level (with sign extension) in both SSE2
and AVX2, and then converting the results back to the 16-bit level
after a right shift.

Bug: aomedia:3455

Change-Id: If8504888cd8a905eb4f550c50305459d64c829c2
(cherry picked from commit 38d838c15bdb64621de89b6dc505607a177eaf86)