Add a USDOT path for av1_warp_affine_neon horizontal filter
The existing implementation widens the inputs to 16 bits each to avoid
problems with one of the inputs being signed and the other being
unsigned. By making use of the USDOT instruction we can avoid this
widening step and therefore process twice as much input data per
instruction.

Benchmarking this with Clang 16 and GCC 12 on a Graviton 3 instance
shows a ~18% reduction in runtime according to the warp affine speed
tests.

Change-Id: I9ab4f25f99bcafc8144833f18e93892a1282d9f5