Add a 16-bit SDOT path for av1_warp_affine_neon vertical filter
The vertical filter performs 16-bit multiplication so cannot use the
Neon SDOT or USDOT instructions which are 8-bit only, however a 16-bit
SDOT instruction is available in SVE.

To use the SVE intrinsics without translating the entire implementation
to SVE we must use the arm_neon_sve_bridge.h intrinsics which are
currently only implemented in Clang. If these are available then we can
insert a Neon vector into the low 128-bits of a SVE vector type and
extract a Neon vector from the low 128-bits of a SVE vector type, which
are both no-ops in the generated assembly.

This patch appears to be beneficial even if the SVE vector length is
greater than 128 bits, even though in these cases we are wasting the
high portion of the SVE vector.

Benchmarking this with Clang 16 on a Neoverse N2 machine shows a geomean
9% reduction in time taken according to the warp affine speed tests.

Change-Id: I6ca236fc7e44385be4f06d311722caa7b0a6956a