rtc: Fixes to RC for SVC and dropped frames
-Delay the use of the clamping in adjust_q_cbr,
based on the superframe counter.
-Add condition that rc->q1/q2_frame > 0
-Replace key_frame condition with frame_is_intra_only(),
as intra_only frames are used in SVC.

This prevents some poor RC behavior where very low Q
was being set early on after a key frame.
This can happen for SVC with intra-only frames on
spatial enhancement layers, or with dropped frames after
the initial key frame.

Change-Id: I28486a82a4e021aca3a921487b7f16b2174d6614