Add an initial Neon implementation for highbd_warp_affine_neon
There are a few more optimizations we could do here, especially for the
p_width==4 cases and specializing based on alpha/beta/gamma/delta, but
let's leave those for another commit to avoid making this individual
change too large.

On a Neoverse N2 machine compiling with Clang-16 and GCC-12 this new
Neon implementation is approximately 3x the speed of the existing C
implementation.

Change-Id: Ia294dce262ca07080f1d8614e98b4c16352bfb2d