Changes to twopass_update_bpm_factor()
Changes to update of the  bpm factor to improve rate accuracy.

The bpm factor (bits per macro block) is used to adjust the
estimate of a suitable Q range for a group of frames.

Updates are limited in magnitude, are damped and are applied after
each ARF group based on the measured rate error for the preceding
group of frames.

Previously the damping was applied to a clamped value. Now the
damping is applied to a raw rate error factor before clamping to the
min/max range for adjustment.

This change whilst keeping the same min max for adjustment for
a single group, also allows the adjustment to compound to a larger
range over multiple groups (except where the rate error tolerance is
set to 100).

The overall impact of this change is to improve rate accuracy for a
significant number of clips including examples that previously
overshot and undershot the target. This comes at the cost of a small
average drop in bdrate performance.

Examples:

Error tolerance set to 25% (Target rate, old rate, new rate)

Bowing (400, 346, 386)
Bridge close (2800, 3390, 3217)
Pamphlet (400, 308, 353)

STATS_CHANGED

Change-Id: I037c700042a95e67f1dc1af23bb1fcd38f1d83a8