Weaken pruning for RESTORE_SWITCHABLE
It is possible for the restoration search functions to find a filter
which is better than RESTORE_NONE when looking purely at SSE, but for
it to be rejected overall due to its rate cost. In this case, there is
a chance that it may be have a lower rate cost when looking at
RESTORE_SWITCHABLE, and so it might be acceptable there.

Change the pruning logic to allow this case through - the time cost
of checking is minimal, and it can allow filters to be used slightly
more often.

Neutral encode time, minor BDRATE improvements:

 Speed | lowres2 | midres2 |  hdres2
-------+---------+---------+---------
   1   | +0.012% | -0.006% | -0.025%
   2   | +0.010% | -0.010% | +0.004%
   3   |  0.000% | -0.005% | -0.002%
   4   | -0.005% | -0.002% | -0.001%

STATS_CHANGED

Change-Id: I257db6a4b3f6069fb0bcd2d2671c35cadd2f27de