Use 4x8 transpose for 4x1 vertical filter in av1_warp_affine_neon
Currently the vertical filter contains some if statements to try and
avoid doing unnecessary work for the p_width == 4 case, however it
appears beneficial to explicitly specialize this further since we can
also use a smaller transpose implementation.

The transpose_4x8 implementation in transpose_neon.h appears to be
unused so this commit also takes the opportunity to clean up the
parameter declarations: the input-only parameters can be passed by value
rather than by pointer.

Benchmarking this on a Neoverse N2 machine with Clang 16 and GCC 12
shows a geomean 3% reduction in runtime for the warp affine speed
tests.

Change-Id: I09f7d76018bcd2181e2943f4ca0f6fa637589c4e