Specialize av1_warp_affine_neon on gamma == 0
Profiling suggests that 0 is a common value for this parameter so it is
worth specializing the implementation here.

In particular for the vertical filter we must currently perform a load
of the filters and also a transpose, however if we know that there is
just a single filter to apply then we can avoid the transpose by making
use of the lane-indexed multiply-add instructions instead.

Benchmarking this on a Neoverse N2 machine with Clang 16 and GCC 12
shows a ~6.3% reduction in the time taken by the warp affine speed
tests.

Change-Id: I2bd22f699b231cbc3b69989c14b0f0aa6569bf18