Disable debugging code in disflow SSE4.1
The SSE4.1 code for disflow contains an internal cross-check against
the C logic at many points. This was added to help with debugging
issues in the SIMD code, but it is very expensive, slowing the SSE4.1
implementation down to the same speed as the C implementation.

This debug code was therefore not meant to be enabled by default,
but was accidentally left enabled. This did not affect release builds,
as the extra code is dead in release mode and gets optimized out,
but it was causing a significant slowdown in debug builds.

Resolve this by disabling the internal cross-check by default.

Change-Id: I90055d807f8063258ffa52655e33d8d0ac0fc80d
(cherry picked from commit d20f66b1c8500ba45e72f1898f9bcd90a2390318)