Remove global-motion-method argument to aomenc
This argument was added for testing purposes, but in practice disflow
is the better option in practically all cases. Therefore this is not
a useful option to expose to end users.

Importantly, we have not done a full release since this option was
added to the codebase, so this argument is not part of the stable
API yet. Therefore it is safe to remove.

Change-Id: I898465904f9a7738f46e47c62f359ba573d47ea5
(cherry picked from commit 0c56d1292056a7e2af6e7f5d1fad018c78291d5f)