Remove w = 2 store paths in Neon convolve_2d_sr horiz pass
The intermediate buffer storing the horizontal-pass output of
convolve_2d_sr is always declared to be the same size - big enough to
hold 128x128 blocks. This means there is no chance of memory over-
writing if storing a block of width 4, when the actual block size is
2x<h>. This patch removes the paths checking if w == 2, storing only
2 elements if true. This removes a branch from a number of hot inner
loops and reduces code size.

Change-Id: Ia24ef747a609444ddbfde8a9efdee6792349b04f