Specialize highbd_warp_affine_neon on gamma == 0
Currently in the vertical filter we have to load the filter values and
additionally perform a transpose. If gamma == 0 then all the filter
values are identical and we can then make use of the lane-indexed
multiply instructions to avoid needing to perform the transpose.

Benchmarking this on a Neoverse N2 machine with Clang 16 and GCC 12
shows a geomean 11.3% reduction in runtime of the warp affine speed
tests.

Change-Id: I259f652e07f413d1d61c6a67f0eede48ed0ee7ff