rtc: Fix to svc patterns in fixed svc mode.
For fixed svc mode: the reference update was
done incorrectly for TL1, which was giving
wrong pattern for 2TL with #SL=2 or 3.

The refresh was only happening on TL1 frame
if #TL=3. But it should also happen if
spatial_layer_id < number_spatial_layers - 1,
as that updated reference slot is then used
for inter-layer prediction on upper spatial layer.

Change-Id: I07a335d9193dee67e55d6e90156e4ce06f734ad1