Avoid saturating narrows in Neon DCTs
The saturation in the definition of custom_packs_s32 appears
unnecessary. Using non-saturating narrowing allows the compiler to
perform the pair of narrows as a single UZP1 instruction.

This change appears to give a geomean 6.7% speedup over the existing
Neon implementation, when benchmarked with Clang 16 and GCC 12 on a
Neoverse V1 machine.

Change-Id: Ice7bfc4d71f540229a80d712cf4fa9eb4d2f5cd3