Move additional transpose functions out of av1_fwd_txfm2d_neon.c
Currently this file contains a few hard-coded transpose functions that
are not specific to that code, so move them to the more general
transpose_neon.h file instead.

We also take this opportunity to fix up the naming scheme of these
functions to match the new transpose_neon.h naming scheme.

Change-Id: I148a3ab56fc91b71fdbee6fafcc4cc38f8aee385