Use narrower vector types in txfm2d 4x8 Neon kernels
The 4x8 kernels here only ever make use of the low half of the Neon
vector that is passed to them, so adjust the interface to these helper
functions to match and propagate changes outwards. This mostly follows
the earlier commit that adjusted the 4x4 kernels.

Also fix the signatures of transpose_arrays_s16_4x8 and
transpose_arrays_s16_8x4 to take int16x4_t parameters to match the
naming and rework the implementations to match, saving a few
instructions:

* For transpose_arrays_s16_4x8 we already only use the low half of the
  vector immediately as the first step in the implementation, so we just
  need to widen it with a zero-cost vcombine to make the intrinsics
  work.

* For transpose_arrays_s16_8x4 we rewrite the implementation to use the
  full vector width up until the final step of the transpose, based on
  the existing 8x8 transpose.

Note that the 16x4 txfm2d implementation could also make use of this
simplification but we do not currently provide a 4x16 fdct/fadst
implementation. Providing these kernels is left for another commit.

Change-Id: Ic33ddb8c68243405c379dd462398df8da49b5d8d