Specialize Neon HBD dist_wtd_sub_pixel_avg_variance
Use the same specialization by filter value as for standard bitdepth.
The rationale for the specialization is as follows:

The optimal implementation of the bilinear interpolation depends on the
filter values being used. For both horizontal and vertical interpolation
this can simplify to just taking the source values, or averaging the
source and reference values - which can be computed more easily than a
bilinear interpolation with arbitrary filter values.

This patch introduces tests to find the most optimal bilinear
interpolation implementation based on the filter values being used.
This new specialization is only used for larger block sizes.

Change-Id: I96b839f4680e64c0cfaad74c6c3af8968f381ced