Rework cospi arrays for av1_fwd_txfm2d_neon.c to avoid shifts
Similar to a previous commit for the sinpi array: the amount that we are
shifting each value loaded from the cospi array is already statically
known based on the index it is loaded from, so adjust the constants to
avoid needing to do a shift at all.

Taking advantage of vectorisation here to load multiple constants at
once is less obvious since not all constants are loaded by all routines,
for example in some routines only indices 16, 32, and 48 are loaded, so
this is left for a later commit.

Change-Id: I567ba1bf5c29c3a87672b2589ac2c294228f21e0