Calculate butterfly pair indices in av1_fwd_txfm2d_neon.c
The indices into the cospi array used in the txfm2d functions are always
symmetric: the second element w1 is always located at `64 - w0_idx`.

This means that for the uses of the butterfly helpers we instead can
provide just a single index and a pointer to the cospi array, and let
the helpers themselves deal with loading the appropriate constants.

Change-Id: I64f31f46054227df31c1f6348addd690924d3691