Rework fadst{4,8}x4_neon and avoid scalar shifts
The amount that we are shifting each value from the sinpi array is
already statically known based on the index it is loaded from, so adjust
the constants to avoid needing to do a shift at all.

Once the shifts are removed we can take advantage that all the values we
need are consecutive in memory and load them all into a single vector,
then make use of the widening lane-indexed multiply instructions.

Change-Id: Icddb1034397021b04c6a697ff216ec29cdb6a5ff