Rework av1_fwd_txfm2d_neon.c btf helpers to load cospi vectors
The access pattern into the cospi array from the butterfly helpers is
easily simplified: we always access `w0_idx` and `64 - w0_idx`, so
rearrange the data layout to put these elements adjacent to each other.

We also observe that the negation of the constants can be done ahead of
time and simply loaded alongside the positive elements.

With these changes we now load all four elements as a single int16x4
vector and then use indexed multiply instructions to avoid needing the
separate broadcast step implied by use of the _n intrinsics as done
previously.

Benchmarking this with Clang 16 and GCC 12 using the txfm2d speed tests
reports a geomean 2% reduction in runtimes.

Change-Id: I07edd10036265667beea18e1600772ac9ad81500