rtc: Adjust condition for int_pro_motion in variance_part
For video (non-screen): in variance_partitioning the
int_pro_motion() was being entered always (if feature is enabled),
whereas for screen it was conditioned on the superblock
source_variance and source sad. These conditions should also apply
to non-screen, which generally gives some speedup with little/no
loss in bdrate.

Also enable the feature of estimate_motion_for_var_based_partition
for SVC, for all speeds, as this gives about 2% speedup at
speed 10 for 2TL on rtc_derf, with ~0.5% bdrate loss.

For single layer video: IC% gain of ~1-2% for speeds 7 and 8 on
rtc_derf and rtc sets (feature is off for speed 9/10), with
small or neutral bdrate change.

Bitexact change for screen.

Change-Id: I8f491582ab996bade699426ecd89de64a06d5e95