Clean up highbd_fwd_txfm_neon.c transposes, remove TRANSPOSE_4X4
We already have a 4x4 transpose definition in transpose_neon.h, so use
that instead and clean up the uses.

We also have an 8x8 transpose in transpose_neon.h however the indices
don't quite match up, so for now just rewrite them in terms of the
existing 4x4 transpose as an initial step in the right direction.

Change-Id: Ie60d15034039e63fa4a190257e3a42e9d51384d7