Fix loop restoration allocs in encoder
In commit c65670f63508d3848442e7159fe1776da77482aa, we changed the
encoder side logic so that it can try multiple different loop
restoration unit sizes. Unfortunately, this change missed one location
where the previous size selection logic was hardcoded, in the code
which allocates worker structs for multithreaded loop restoration.
This led to an out-of-bounds write under certain conditions.

Fix by fetching the minimum loop restoration size to search, as this
value sets the upper limit on the number of workers needed.

Bug: aomedia:3506
Change-Id: Ib72d118a5719af6789a6e073d943d98186d5f97b