Remove unused UNINITIALIZED_IS_SAFE macro
This macro does not have a definition on MSVC and we have already
avoided all such uses anyway, so this can just be removed completely.

There is one outstanding use of the macro in loopfilter_neon.c, however
this appears to be pointless since the variables are already fully
initialized so we just remove it.

Bug: aomedia:3507
Change-Id: I6a8db44c7c0f09bcaeacd8b5399ab8aa59721da9