Remove partial vector loads from mem_neon.h
These were only ever used to either (a) load the 0th lane of the vector
(in which case we can simply zero-init the rest at zero cost), or (b)
used as one of multiple calls to fully initialise the vector with a
stride. The use of a macro with immediate is problematic since it
depends on the compiler to be able to fully analyze the use of the
vector to determine whether the initialization is safe or not.

For loading the 0th lane we already had a `load_u8_4x1_lane0` to handle
the zeroing automatically which we now use everywhere where it is
needed, and for fully-initializing the vector with a stride we can
simply pass in the stride and do the loading in the mem_neon.h helper
function. This has the nice advantage that we can remove the load macros
entirely.

Additionally, since the `_lane0` suffix is no longer needed to
distinguish from the lane-immediate macro version, the suffix is
removed.

Bug: aomedia:3507
Change-Id: I33ea4885c96c96059f0de67a5fdc4024b588d228