Change AV1E_SET_MAX_CONSEC_FRAME_DROP_CBR to int
Change the parameter type of AV1E_SET_MAX_CONSEC_FRAME_DROP_CBR from
unsigned int to int because this value is stored in a struct member of
the int type. Change the codec control to return AOM_CODEC_INVALID_PARAM
if the value is negative.

Document the parameter type in the comment.

Note: A follow-up to
https: //aomedia-review.googlesource.com/c/aom/+/182921.
Change-Id: I99e6180358c1cdf72cf7a9ad32a138e784c8d461