Rework upscaling in disflow algorithm
Previously, when upscaling the flow field from one pyramid level to
another, we used the same upsampling filter as for upscaling pixels.

However, this has a couple of issues: the pixel filtering code is
much more generic than we need for flow fields, because flow fields
are always upscaled by a factor of exactly 2; and we were not
correctly accounting for the offset between the flow vector locations
at one pyramid level and the next pyramid level down.

Therefore, replace this logic with a much simpler filter, which only
handles 2:1 upscaling and which correctly accounts for the relative
positions of the input and output grids.

After this replacement was made, four filters were tested: Bilinear,
bicubic, and 6- and 8-tap fullband filters taken from elsewhere in
the code. Bicubic performed slightly better than the rest, so use
this option.

This change simplifies the code and provides a better starting point
for SIMD implementation, as well as giving minor speed and quality
improvements.

 Speed | BDRATE-PSNR | BDRATE-SSIM |   Enc time
-------+-------------+-------------+-------------
   1   |   -0.004%   |   +0.001%   |   +0.006%
   2   |   -0.014%   |   -0.003%   |   -0.021%
   3   |   -0.028%   |   -0.073%   |   -0.034%
   4   |   -0.009%   |   -0.003%   |   -0.101%

STATS_CHANGED

Change-Id: I040b486be370627bdb68fbfd57ba855c01c2ea72