rtc: Set non_reference_frame flag off on key frames.
This is guard against scenario where the user sets
the prediction pattern with SET_SVC_REF_FRAME_CONFIG,
and on a key frame sets the refresh flag off for all
buffer slots off/0. If so, then that key frame can have
rtc->non_reference_frame set to 1 and avoid loopfilter,
which is incorrect. So this fix prevents this from happening.

Change-Id: Ie426d8acc39a07f40473a166be02762ed5d284eb