mem_neon.h: Introduce and use strided store helpers
Several of the z3 predictor implementations make use of consecutive
store-lane instruction to scatter parts of the vector to strided
locations in memory. To save this duplication, add a new helper to
mem_neon.h for handling this strided-scatter operation.

It turns out that similar helpers already exist, but the naming scheme
currently clashes with the multi-vector store cases.

Remove all uses of the following macros, since they were either
never used or are can now be calls to the strided store helpers instead:

* store_s16_2x1(s, s0, lane)
* store_u16_2x1(s, s0, lane)
* store_u16q_2x1(s, s0, lane)
* store_unaligned_u8_4x1(dst, src, lane)
* store_unaligned_u8_2x1(dst, src, lane)
* store_unaligned_u16_2x1(dst, src, lane)
* store_unaligned_u16_4x1(dst, src, lane)

There is one remaining (non-strided store) use of store_u16_2x1, however
this can now be made into a function since it is only ever using the low
lane of the vector.

With this change the only macros exposed from mem_neon.h are now
store_u8_2x1 and store_u8_4x1. These are left for a later commit since
they have many more existing uses, primarily for the convolve kernels.

Change-Id: I59843f54e3a443fa8eefb3e6bf78baa68b3698b0