warp_plane_neon.h: Avoid loading all data separately from filter
Currently the warp affine horizontal filter kernel spends a lot of time
loading data only to then store it back to the stack due to the need to
check the boundary condition before applying the filter. This was
originally done to avoid duplication of the loop condition, however it
is beneficial to apply the filter immediately to avoid the pointless
memory access.

Benchmarking on a Neoverse N2 machine with Clang 16 and GCC 12, the
speed tests show a geomean 7.3% reduction in times reported from this
change.

Change-Id: Ib7cbcf23acdc2ed8b3af4cd9df9044d98c47aff0