Fix potential memory leaks in av1_loop_restoration_alloc()
If the allocation of 'rst_tmpbuf' or 'rlbs' fails in
av1_loop_restoration_alloc(), 'lr_sync->num_workers' which is used
to free these buffers will not be set due to longjmp. This will
cause the allocated buffers to show up as leaks. This CL fixes this
issue by setting 'lr_sync->num_workers' before the allocation of
buffers. Also used aom_calloc() to allocate 'lrworkerdata'
to avoid the SEGV error while trying to free the unallocated buffers.

Bug: aomedia:3276
Change-Id: I0dca677e13428b1df9599f953d4308d2dfaea301