Fix a potential SEGV error in av1_terminate_workers()
In the parent commit, workers are created in av1_create_workers()
in descending order, and terminated in av1_terminate_workers()
in ascending order. If creation of a worker (other than first and
last worker) fails due to an allocation failure, this will cause a
SEGV error while terminating workers who were not created (workers
are allocated using aom_malloc()). This CL fixes this issue by
creating workers in ascending order.

Bug: aomedia:3276
Change-Id: Iaefba5917d486acc82df6d01bbab2d262e2e1801