intrapred_neon.c: More perf improvements for z2 predictor
The existing `shuffle` helper functions serve to move the loaded data
into the high lanes of the result and zero out the low lanes for the
remainder, however this is unnecessary since the zeroed data is already
masked out by the later BSL instructions that we use to merge the
calculations from the `left` and `above` inputs. This means we can
instead just use a normal unmasked load instruction like we already do
in the highbd implementation.

To avoid duplication we extract the individual `above` and `left`
calculations into their own helper functions. This then allows us to
provide dedicated paths for the case that only one helper is needed,
avoiding a pointless merge with an empty vector.

Also replace compare into select with max: `a > b ? a : b` is just
`max(a, b)`, and avoid computing `min_base_y >> 1` in the loop body.

Benchmarking on a Neoverse N2 machine with Clang 17, this change shows a
geomean 34.35% reduction in times reported.

Change-Id: I321f2f462dbf706b9832ce07bbb45332b76c99de