Fix the error handling in av1_tf_info_alloc()
An allocation failure in av1_tf_info_alloc() is reported using
'cm->error'. This can result in a SEGV error when the function is
invoked by av1_check_initial_width() call in encoder_encode()
as 'cm->error.setjmp' is not set at this point. This CL propagates
the error status from av1_tf_info_alloc() to the caller of
av1_check_initial_width() so that the appropriate 'error_info'
can be used for error reporting.

Bug: aomedia:3276
Change-Id: I8bdd570584378abe4fcebaf95731909cddb8c1e1