Add back the if statement in set_size_literal()
Add back the if statement that checks whether we need to reallocate
compressor data in set_size_literal().

We are not sure that if statement will never be executed. So it is safer
to add it back.

Bug: aomedia:3526
Change-Id: I9953da3f218571e7eeca1b295ba7791fb03a5082
(cherry picked from commit 4f632e5b6d1c3cce0bbbc0f2c030171082fc1b3e)