Assert intra mode in av1_predict_intra_block()
The assertion is intended to suppress Coverity CID 340240:
Out-of-bounds access (OVERRUN). Coverity thinks that extend_modes[mode]
in av1_predict_intra_block() may be an out-of-bounds access, which is
incorrect if `mode` is an intra mode.

Also explicitly declare the size of the mode_to_angle_map array as
INTRA_MODES.

Change-Id: I403c7deb61a418da993169b05b523b78ffcb19b3