dr_prediction_test.cc: Allocate 80 * bh buffer for tests
Rather than reusing the same 64 * 64 buffer, since this should enable
address-sanitizer to give much more useful errors if we accidentally
write out of bounds.

We only do this for the output buffer rather than the left/above input
vectors for now, since the latter appears to be okay with partially
over-reading the input and allowing this is useful for performance.

Change-Id: I86f6d297625ab56970dae7bc5db31c0818487b8e