Optimize av1_predict_intra_block() for higher bit-depths
For HBD encode/decode, pred data for intra modes was generated using
the functions av1_predict_intra_block() and
highbd_build_intra_predictors(), in parent version. The computations
in these functions were generalized for all intra modes, and
evaluation of some of the intra modes can be optimized based on
neighbor availability and mode information. Since non-directional
intra modes require only left and top neighbors, this CL avoids the
redundant checks for these modes. Additionally, computations of
filter intra and edge filter type are avoided as the same are not
required for these modes.

The optimizations in this CL are similar to the LBD path, as in the
following CL:
https://aomedia-review.googlesource.com/c/aom/+/185801.

                 Encode Time
cpu-used         Reduction(%)
    6               1.328
    7               2.049
    8               1.495
    9               1.007

Bug: aomedia:3532
Change-Id: I9bf7d5036f9de1f15781e2364e4aca387ea9cf90