Make img_alloc_helper() fail on AOM_IMG_FMT_NONE
If fmt is AOM_IMG_FMT_NONE, currently img_alloc_helper() allocates a
single plane because AOM_IMG_FMT_NONE (0) is not a planar format (the
AOM_IMG_FMT_PLANAR bit is not set in AOM_IMG_FMT_NONE).

Although this seems correct, the problem is that most of the code in
libaom assumes planar formats and is likely to dereference a null
pointer when it uses img->planes[1]. Also, AOM_IMG_FMT_NONE isn't really
a valid image format. So it is safer to make img_alloc_helper() fail if
fmt is AOM_IMG_FMT_NONE.

This is the libaom version of the libvpx CL
https://chromium-review.googlesource.com/c/webm/libvpx/+/5215311.

Change-Id: I40869b4e694c907dbe7667300cd46b8754cb7440