Zero initialize an array in cdef search
This array was not initialized.

The array is supposed to store the best corresponding cdef filter
strength. And its value is an index of another array.

Without the initialization, this might lead to out of bound access
to the other array.

Change-Id: Id65c448f479865992ada0b09ff6e97b2d2ac3362
(cherry picked from commit 6fcad835fbd1cba379f34ecb96755888d21a98df)