Disable the check on 0 target_bitrate in validate_config
This is causing failures in latest libaom roll here:
https://webrtc-review.googlesource.com/c/src/+/337284

That check was added as part of this CL:
https://aomedia-review.googlesource.com/c/aom/+/186601

Revert that part for now to enable roll to proceed.

Some webrtc tests are creating the encoder with 0 target_bitrate
and then assigning non-zero bitrate before encoding the first
frames, so the check in validate_config() was causing
InitEncode to fail.

Bug: aomedia:3523

Change-Id: I8462ffba1d80219272d6225e572e047184337cd7