Document frame size to copy in loop filter
In AV1, loop filter applies to the whole frame according to the
aligned with and height.

Comments are added to resolve a TODO and make it clear to understand
the filtering process of loop filter and super res.

Comments are also added in the buffer struct for y_width and
y_crop_width.

Change-Id: Idd80603f5b840421237d3edd4829212f585841d9