mem_neon.h: Define vld1q_u16_x4 until GCC 8.5.0
When trying to build the library with GCC 8.3.0 the vld1q_u16_x4
intrinsic is not available, leading to the code failing to compile.

We already provide a definition for this helper for earlier versions of
GCC, so adjust the checks so we also provide it until GCC 8.5.0 instead.

Bug: aomedia:3543
Change-Id: I98a32ae6abd068f326a1075dd6782b190e2eac1d
(cherry picked from commit 0b2a8639c2aef7494d333fae47b3817d3c08e3d4)