Add new global motion speed feature
Add a speed feature which generates the feature list used
in global motion estimation to be generated from a downsampled
version of the input frame.

Tuning:
* downsample_level = 0 for speed <= 3
* downsample_level = 1 for speed 4,5
* downsample_level = 2 for speed 6

Relative to the previous patch *with global motion forced on
for speed 5 and 6*, the impact of this change is:

 Speed | BDRATE-PSNR | BDRATE-SSIM |   Enc time
-------+-------------+-------------+-------------
   4   |   +0.032%   |   +0.016%   |   -0.762%
   5   |   +0.053%   |   +0.051%   |   -1.569%
   6   |   +0.272%   |   +0.286%   |   -3.242%

IMPORTANT: Global motion is not enabled yet on speed 5 or 6,
so only speed 4 is impacted right now. However, this patch is
a step toward enabling global motion for higher speeds.

STATS_CHANGED

Change-Id: If0a1640c6f24b9bf91c7350b9a905c9d7f673695