Add SVE2 implementation of HBD convolve_x_sr for 4-tap filters
Add SVE2 implementation of av1_highbd_convolve_x_sr for 4-tap filters as
well as the corresponding tests.

This implementation is not beneficial on cores that only support SVE
(like Neoverse V1) due to them having 256-bit SVE pipes instead of
128-bit pipes on more modern cores (that also support SVE2). It
therefore makes sense to restrict this implementation to SVE2 only. This
also allows us to use extra instructions not available in SVE.

Change-Id: I06590e03b56a262f01a25324fc2d2ebb88369495