rtc: Fix to setting content state for active_map
When a coding block has the segment skip set
(i.e, inactive segment in active maps) the content
state should be zero sad. This change fixes this setting
to only set zero sad for the coding block and not the
whole superblock. This reduces artifacts observed
in active maps on boundary regions (superblocks with
part of it labelled as inactive).

Change-Id: I682966394391bdf7d25c48522f37c475b9e35a74