Specialize 8-tap HBD dist_wtd_convolve_y_sve2 on bitdepth
The rounding value at the end of the convolution depends on the bitdepth
(8/10, or 12). Add 2 specialized versions of the function, so that we
know the rounding value at compile time and therefore use only one
instruction instead of two to perform the final rounding and narrowing
step. This gives up to 10% uplift over the non-specialized version.

Change-Id: Id30c3ba63567ff92ff0a79aaf94c091e0d807f02