Move force_zeromv_skip logic up in nonrd_pickmode
Move the skip logic for force_zeromv_for_blk further up
in the nonrd_pickmode, to avoid possibel conflict with
the sb_me_block feature, which has an early exit that
conflict with force_zeromv_skip. Currently this has no
effect, but change it now in case the sb_me_block feature
is modified.

Bitexact on speed 6-11 for screen, with neutral
IC speedup.

Change-Id: I94160ac6c6dd2b7a4025c20bcf62d2c9e97eff65