rtc: Move force_zeromv_skip logic up in nonrd_pickmode
Move the skip logic for force_zeromv_for_blk further up
in the nonrd_pickmode, to avoid possible conflict with
the sb_me_block feature, which has an early exit that
conflict with force_zeromv_skip. Currently this has no
effect, but change it now in case the sb_me_block feature
is modified.

We also need to move up the check on use_ref_frame_mask,
otherwise the frame_mv (in the force_zeromv_skip_for_blk check)
may not be set. This is a re-application of the patch
https://aomedia-review.googlesource.com/c/aom/+/188741, with
the fix for the valgrind issue.

Bitexact on screen, with neutral IC speedup.

Change-Id: I33d4fcb25022abf43a109e7412cf7ea6b12f053b