Remove no longer needed special case in Neon convolutions
The neon_dotprod and neon_i8mm implementations of some convolution
functions have a special case to handle the no-op filter, as 128 does
not fit in a signed 8-bit integer. This no-op filter is never used in
practice - the unit tests have been updated to reflect this in a
previous patch - so remove the code to handle the no-op filter.

Change-Id: Ifd43a4f752a6cb9de8bbfbe7b4577764a4208d7e