Add 6-tap filter path for convolve_vert_scale
The filter values used in the scaling algorithm are specified in the
documentation of the inter-prediction process (chapter 7.11.3.4. [1]).
These filter values are defined in the av1_interp_filter_params_list
in filter.h. An important characteristic of these filters, except the
MULTITAP_SHARP filter, is that at indices 0 and 7 the values are 0.

Add an implementation for vertical filtering that specialises on
6-tap filters. This way we avoid redundant work associated with 8-tap
filters.

[1]https://aomediacodec.github.io/av1-spec/av1-spec.pdf

Change-Id: Ib2863cd29541a3d69d128cce9da0485de8ad182f