Remove MSVC version check in x86 workaround
The MSVC version check is not future-proof unless the bug has been
fixed. Since we have not even submitted a bug report, this bug is
unlikely to be fixed soon. So remove the MSVC version check for the
workaround.

Update the comment to note that this bug can still be reproduced with
Visual Studio 2022 version 17.10.3.

Bug: 349646650
Change-Id: If2935164d20a52b0c7daf7f88ea8ae588e2f6efa