Fix to QP for temporal enhancment after keyframe
For temporal enhancement layers in RTC mode:
after a key frame use the avg_qindex_key for the
ambient_qp for the active_worst_quality setting.
And use last_q[KEY/INTER_FRAME] depending on keyframe
distance in adjust_q_cbr().

This reduces some sawtooth-like behavoior in psnr/qp
observed for temporal layers with periodic keyframes.

This also exposed a bug in the external RC where
frames_since_key was not updated in the correct place.

Change-Id: Ia63819a98823b13aefdde79b934fb30b78953986