Do not add -lm to Libs.private if WIN32 or APPLE
In Windows and Apple operating systems, the functions declared in
<math.h> are defined in the main standard C library rather than a
separate libm. Do not add -lm to the Libs.private field in aom.pc if
WIN32 or APPLE is true.

This condition matches the condition under which we link the aom and
aom_av1_rc libraries with -lm in CMakeLists.txt.

Bug: 356153293
Change-Id: Ic42c84b0583aafa4491e393d691939c293415986