Refactor 4-tap convolve_2d_sr Neon I8MM path
A 6-tap USMMLA (horizontal) convolution kernel requires the same
number of instructions as a 4-tap USDOT convolution kernel; therefore
we can use the USMMLA 6-tap path for both 6- and 4-tap cases.

This patch uses the above information to expand the utility of the
Neon I8MM 4-tap combined 2D convolution to support up to 6-tap
horizontal filters. This is useful because 6-tap horiz, 4-tap vert
filter combinations are the third most common type for convolve_2d_sr
in --rt encodings after 6, 6 and 4, 4.

Change-Id: Iefaf3a4f759bbcfd61ff60c829225d4aff93556c