Clean up av1_resize_and_extend_frame_neon
Clean up implementation in order to reuse code when adding
implementations for different architecture extensions in subsequent
patches.

- Move horizontal and vertical filtering into their own functions.
- Move widening into the main loops from scale_filter_8 function and
reuse vectors in the next loop iteration.
- Use non-rounding shifts in the bilinear kernel which are generally
faster than rounding shifts on modern CPUs.

The refactoring was done for the implementation of 2 to 1 and 4 to 1
scaling only because of a bug[1] in the implementation of 4 to 3
scaling.

[1]Bug: 363916152

Change-Id: I3fbf6cc0731c7d4889ae935f44c6f900cd5f1550