Add I8MM 2/1 scale spec. for av1_resize_and_extend_frame
Add an Armv8.6 I8MM implementation for av1_resize_and_extend_frame
that specialises on 2 to 1 scaling. When BILINEAR filter is used or
phase value is equal to 0 the Armv8.0 Neon implementation is called.

Use the USMMLA matrix multiply instructions to accelerate the
horizontal pass, utilizing the fact that both the EIGHTTAP_SMOOTH and
EIGHTTAP_REGULAR filters are 0-padded to 8 taps. (USMMLA does twice
the work of a USDOT dot product instruction.)

Change-Id: I8965a6747840c7b59486d8fed537218d629648cd