arm/warp_plane_*: Hoist filter load if alpha == beta == 0
In warp affine the common case is that both alpha and beta are zero, in
which case the vector of filter values does not need to be re-loaded on
each iteration of the loop.

LLVM 18 appears to struggle to do this itself, so manually hoist out the
vector and pass it as a parameter instead.

On a Neoverse V2 machine this seems to give a geomean ~0.9% reduction in
times measured for the warp affine speed tests for the cases where alpha
and beta are both zero.

Change-Id: I1eaffe07e55f8c93f1d4a822321cbe09f8250c99