Optimize av1_apply_temporal_filter_neon
To perform filtering we have to calculate sum of elements which are
part of a 5x5 size window. This window is a sliding window, which
means that in a new loop cycle we are calculating the sum of 4 rows
from the loop cycle one before and a newly loaded row. Instead of
recalculating the sum at every loop cycle, subtract the sum of
elements of the row which is not part of the sliding window anymore
and add the sum of elements of the newly loaded row.

Vectorise the addition between the sum of elements of the sliding
windows and luma_sse_sum.

Change-Id: I87c04cb8c6b9bfa73319c2ca783e50feafe1f4ee