metadata_test: fix test w/CONFIG_REALTIME_ONLY=1
after:
278c942366 aom/exports_test: rm aom_*_metadata_to_frame_buffer

The test was removed in test.cmake, but relied on the file being added
first. In the breaking change the file was only added afterward, when
BUILD_SHARED_LIBS=0.

In this change keep metadata_test.cc in the source list for
CONFIG_REALTIME_ONLY=1, but don't declare MetadataEncodeTest which
requires the one pass good encoder.

Bug: aomedia:370599806
Change-Id: I81d06b173a43c8314e19360ef01f88539fc41e57