Remove experimental feature AOM_CODEC_USE_PRESET
The AOM_CODEC_USE_PRESET experimental feature is really usage-specific
default values for extra_config. It is overly restrictive for the preset
feature because each usage is allowed to have only one preset. So we are
removing AOM_CODEC_USE_PRESET and will redesign the preset feature to be
more flexible.

This CL is the minimum change to remove the experimental feature
AOM_CODEC_USE_PRESET from the public API. It is intended to be
cherry-picked to the minty branch for the libaom v3.11.0 release.
This CL essentially reverts commit f10e098.

The rest of the AOM_CODEC_USE_PRESET code will be removed later.

Bug: 359341194
Change-Id: I74fece5a66e3fc262a78e3edcf52bf0356ac2e9d