Don't check size limit in aom_realloc_frame_buffer
DECODE_WIDTH_LIMIT and DECODE_HEIGHT_LIMIT (used when CONFIG_SIZE_LIMIT
is defined as 1) are intended for the decoder only. Since
aom_realloc_frame_buffer() is shared by the decoder and the encoder, it
should not check DECODE_WIDTH_LIMIT and DECODE_HEIGHT_LIMIT. Move the
check to the aom_realloc_frame_buffer() call in the
read_uncompressed_header() function.

Bug: 385367183
Change-Id: Icdb3660e3631fefb690ef9f28a4d782e4c9a939d