test_aom_rc: work around multiple def link error w/cygwin/mingw
When building with `BUILD_SHARED_LIBS=1` the aom_av1_rc library will
link against a static version of libaom as it requires some internal
functions. This causes the library to contain duplicate versions of some
of the symbols in libaom which causes link errors when targeting cygwin
or msys2/mingw-w64. `--allow-multiple-definition` is added to the linker
flags as a workaround. A better fix will be to limit the symbols
exported by the library to only the public interface, as is done with
libaom.

Fixes errors of the form:
.../x86_64-pc-cygwin/bin/ld:
  libaom.a(aom_codec.c.o):aom_codec.c:(.text+0x30): multiple definition
  of 'aom_codec_iface_name';
  libaom_av1_rc.dll.a(cygaom_av1_rc_dll_d000055.o):(.text+0x0): first
  defined here

Bug: aomedia:391715078
Change-Id: Ib6c7431de6b7d6f867371c9518ac4b5fdd80667d