rtc-svc: Fix to the layer_id under dynamic change.
This is a fix to the revert issue here:
https://chromium-review.googlesource.com/c/chromium/src/+/6442206

The failing SVC test changes the svc->number_temporal_layers
dynamically from 2 to 1, and when it switches to 1 it does
not use the ctrl_set_layer_id to set temporal_layer_id to 0.
Since the switch happened on a TL1 frame, the new temporal_layer_id
in the encoder was kept to 1, even though the number_temporal_layers
is 1.

This was causing the assert failure in libaom:
assert(IMPLIES(!has_nonzero_operating_point_idc, obu_extension == 0));

The fix is to clamp the spatial/temporal_layer_id to proper values
on dynamic change of layers.

Change-Id: I36058fd309cbbaee18d59708606a299408983f34