Fix to av1_select_sb_size for spatial layers.
The existing condition in av1_select_sb_size() for
spatial layers was causing test failures when the user passes
in downscaled frames (i.e., externally resized). Tests were
added in the CL below to capture the issue, and there the
superblock_size was set to 64x64 in the test via the
encoder control in order to make the tests pass:
https://aomedia-review.googlesource.com/c/aom/+/199141

Fix here is to constrain for sb_size to be 64x64
in select_sb_size() for spatial layers.

This allows for external scaling to work for spatial layers
with --sb-size=0 (default mode), and fixes the test failures
in: https://webrtc-review.googlesource.com/c/src/+/379680.

Change-Id: I80d32732df925cfd8a4be3a2fa3365be01f158fb