rtc: Fixes for ROI
Allow for ROI delta_qp to be applied on
key frames. And clamp the delta_qp so that
the qindex is above zero on the segment, to avoid
the lossless segment case. Lossless segment can be
triggered in the roi tests by setting min_quantizer
to 0 and using a lower max_quantizer.

Change-Id: I478f00153dee464a12d4b7a9271002452477543d