aom_image: add allocation guidance for aom_img_wrap
This change is the same as the one made in libvpx with 'vpx_' replaced
with 'aom_':
8d00aca6 vpx_image: add allocation guidance for vpx_img_wrap
The image data buffer provided to `vpx_img_wrap()` must have a minimum
allocation size that satisfies the requirements of the `fmt`, `d_w`,
`d_h`` and `stride_align` parameters. This change adds a description of
how to calculate the size based on the one done in `img_alloc_helper()`.

Bug: 442142951
Change-Id: I8d282623ca697ab42ebec9e08efaf3b0dc4a64e3