rtc-screen: Fix to improve scroll quality
For realtime screen: turn off the speed feature
that reduces subpel motion search
(set sf->reduce_mv_pel_precision_highmotion to 0).

And fix the counter updates of the last num_col/row_blscroll,
with adjustment of some conditions for setting the high
motion flag.

These changes give overall better scroll detection for
high resolution screen. In particular it reduces/removes
the visual regression observed in scaled source with scrolling,
observed in the issue below. And it helps to reduce the psnr
dips observed for 2/3x and 3/4x scaling with scroll.

Bug: b:423804955, b:444116924
Change-Id: I1b2e82fd765a1311e5723af2c863d5680a8eb18f