rtc: Adjust psnr to fix nightly rt test failures
Adjust the psnr threshold in rt_end_to_end_test.cc
to fix the nightly failures. This was caused by
turning off the speed feature in:
https://aomedia-review.googlesource.com/c/aom/+/205381

This only affects speed 6 and that speed feature will be
re-enabled if it is not the cause of the crash,
but for now adjust the psnr thresholds for speed 6
(and 7 in some cases to make it consistent).

Change-Id: I3e2a426ab0d5f889e726074fd08deddf5af06ba7